refactor: utils로 파일 이동 및 이름 변경
